feat: adiciona filtro de status e separa filtros das transações
